+++ apps/backend/src/modules/points-of-interest/points-of-interest.service.test.ts
@@ -305,6 +305,7 @@ describe("PointsOfInterestService", () => {
 			routes: [
 				{
 					categories: [],
+					createdAt: "2025-08-15T00:00:00Z",
 					createdByUserId: 0,
 					description: "",
 					distance: 0,