+++ apps/backend/src/modules/routes/route.entity.ts
@@ -14,6 +14,8 @@ type SavedUserRoute = {
 class RouteEntity implements Entity {
 	private categories: null | ReturnType<CategoryEntity["toObject"]>[];
 
+	private createdAt: null | string;
+
 	private createdByUserId: number;
 
 	private description: null | string;
@@ -40,6 +42,7 @@ class RouteEntity implements Entity {
 
 	private constructor({
 		categories = [],
+		createdAt,
 		createdByUserId,
 		description,
 		distance,
@@ -52,6 +55,7 @@ class RouteEntity implements Entity {
 		savedUserRoute,
 	}: {
 		categories?: ReturnType<CategoryEntity["toObject"]>[];
+		createdAt: null | string;
 		createdByUserId: number;
 		description: null | string;
 		distance: number;
@@ -68,6 +72,7 @@ class RouteEntity implements Entity {
 		savedUserRoute: null | SavedUserRoute[];
 	}) {
 		this.categories = categories;
+		this.createdAt = createdAt;
 		this.id = id;
 		this.distance = distance;
 		this.duration = duration;
@@ -81,6 +86,7 @@ class RouteEntity implements Entity {
 	}
 
 	public static initialize(data: {
+		createdAt: string;
 		createdByUserId: number;
 		description: null | string;
 		distance: number;
@@ -97,6 +103,7 @@ class RouteEntity implements Entity {
 		savedUserRoute?: SavedUserRoute[];
 	}): RouteEntity {
 		return new RouteEntity({
+			createdAt: data.createdAt,
 			createdByUserId: data.createdByUserId,
 			description: data.description,
 			distance: data.distance,
@@ -111,6 +118,7 @@ class RouteEntity implements Entity {
 	}
 
 	public static initializeList({
+		createdAt,
 		createdByUserId,
 		distance,
 		duration,
@@ -120,6 +128,7 @@ class RouteEntity implements Entity {
 		name,
 		pois,
 	}: {
+		createdAt: string;
 		createdByUserId: number;
 		distance: number;
 		duration: number;
@@ -134,6 +143,7 @@ class RouteEntity implements Entity {
 		}[];
 	}): RouteEntity {
 		return new RouteEntity({
+			createdAt,
 			createdByUserId,
 			description: null,
 			distance,
@@ -170,6 +180,7 @@ class RouteEntity implements Entity {
 		savedUserRoute?: SavedUserRoute[];
 	}): RouteEntity {
 		return new RouteEntity({
+			createdAt: null,
 			createdByUserId,
 			description,
 			distance,
@@ -202,6 +213,7 @@ class RouteEntity implements Entity {
 	}): RouteEntity {
 		return new RouteEntity({
 			categories: data.categories,
+			createdAt: null,
 			createdByUserId: data.createdByUserId,
 			description: data.description,
 			distance: data.distance,
@@ -217,6 +229,7 @@ class RouteEntity implements Entity {
 
 	public toDetailsObject(): {
 		categories: ReturnType<CategoryEntity["toObject"]>[];
+		createdAt: string;
 		createdByUserId: number;
 		description: null | string;
 		distance: number;
@@ -234,6 +247,7 @@ class RouteEntity implements Entity {
 	} {
 		return {
 			categories: this.categories as ReturnType<CategoryEntity["toObject"]>[],
+			createdAt: this.createdAt as string,
 			createdByUserId: this.createdByUserId,
 			description: this.description,
 			distance: this.distance,
@@ -252,6 +266,7 @@ class RouteEntity implements Entity {
 	}
 
 	public toListObject(): {
+		createdAt: string;
 		createdByUserId: number;
 		distance: number;
 		duration: number;
@@ -266,6 +281,7 @@ class RouteEntity implements Entity {
 		}[];
 	} {
 		return {
+			createdAt: this.createdAt as string,
 			createdByUserId: this.createdByUserId,
 			distance: this.distance,
 			duration: this.duration,
@@ -306,6 +322,7 @@ class RouteEntity implements Entity {
 
 	public toObject(): {
 		categories: ReturnType<CategoryEntity["toObject"]>[];
+		createdAt: string;
 		createdByUserId: number;
 		description: null | string;
 		distance: number;
@@ -322,6 +339,7 @@ class RouteEntity implements Entity {
 	} {
 		return {
 			categories: this.categories as ReturnType<CategoryEntity["toObject"]>[],
+			createdAt: this.createdAt as string,
 			createdByUserId: this.createdByUserId,
 			description: this.description,
 			distance: this.distance,