+++ apps/backend/src/modules/routes/route.model.ts
@@ -2,7 +2,10 @@ import { type RouteUploadImageResponseDto } from "@smartscapes/shared";
 import { Model, type QueryBuilder, type RelationMappings } from "objection";
 
 import { FileFolderName } from "~/libs/enums/enums.js";
-import { DatabaseTableName } from "~/libs/modules/database/database.js";
+import {
+	AbstractModel,
+	DatabaseTableName,
+} from "~/libs/modules/database/database.js";
 import { type LineStringGeometry } from "~/libs/types/types.js";
 
 import { CategoryModel } from "../categories/category.model.js";
@@ -11,7 +14,7 @@ import { PointsOfInterestModel } from "../points-of-interest/points-of-interest.
 import { UserRouteModel } from "../user-routes/user-route.model.js";
 import { type UserRouteStatusType } from "./libs/types/types.js";
 
-class RouteModel extends Model {
+class RouteModel extends AbstractModel {
 	public static override get tableName(): string {
 		return DatabaseTableName.ROUTES;
 	}
@@ -27,8 +30,6 @@ class RouteModel extends Model {
 	public duration!: number;
 
 	public geometry!: LineStringGeometry;
-
-	public id!: number;
 
 	public images!: RouteUploadImageResponseDto[];