+++ apps/frontend/src/modules/categories/categories-api.ts
@@ -2,9 +2,14 @@ import { APIPath, ContentType } from "~/libs/enums/enums.js";
 import { BaseHTTPApi } from "~/libs/modules/api/api.js";
 import { type HTTP } from "~/libs/modules/http/http.js";
 import { type Storage } from "~/libs/modules/storage/storage.js";
-import { type APIResponse } from "~/libs/types/types.js";
+import { type APIResponse, type PaginationMeta } from "~/libs/types/types.js";
 
-import { type CategoryGetAllItemResponseDto } from "./categories.js";
+import { CategoriesApiPath } from "./libs/enums/enums.js";
+import {
+	type CategoryCreateRequestDto,
+	type CategoryGetAllItemResponseDto,
+	type CategoryGetAllQuery,
+} from "./libs/types/types.js";
 
 type Constructor = {
 	baseUrl: string;
@@ -17,13 +22,31 @@ class CategoriesApi extends BaseHTTPApi {
 		super({ baseUrl, http, path: APIPath.CATEGORIES, storage });
 	}
 
-	public async getAll(): Promise<APIResponse<CategoryGetAllItemResponseDto[]>> {
+	public async create(
+		payload: CategoryCreateRequestDto,
+	): Promise<APIResponse<CategoryGetAllItemResponseDto>> {
 		const response = await this.load<
-			APIResponse<CategoryGetAllItemResponseDto[]>
-		>(this.getFullEndpoint("/", {}), {
+			APIResponse<CategoryGetAllItemResponseDto>
+		>(this.getFullEndpoint(CategoriesApiPath.ROOT, {}), {
 			contentType: ContentType.JSON,
-			hasAuth: false,
+			hasAuth: true,
+			method: "POST",
+			payload: JSON.stringify(payload),
+		});
+
+		return await response.json();
+	}
+
+	public async getAll(
+		payload: CategoryGetAllQuery | undefined,
+	): Promise<APIResponse<CategoryGetAllItemResponseDto[], PaginationMeta>> {
+		const response = await this.load<
+			APIResponse<CategoryGetAllItemResponseDto[], PaginationMeta>
+		>(this.getFullEndpoint(CategoriesApiPath.ROOT, {}), {
+			contentType: ContentType.JSON,
+			hasAuth: true,
 			method: "GET",
+			query: payload,
 		});
 
 		return await response.json();