+++ packages/shared/src/modules/categories/libs/types/category-get-all-item-response-dto.type.ts
@@ -1,4 +1,5 @@
 type CategoryGetAllItemResponseDto = {
+	createdAt: string;
 	id: number;
 	key: string;
 	name: string;