+++ packages/shared/src/modules/categories/libs/types/types.ts
@@ -1,2 +1,3 @@
 export { type CategoryCreateRequestDto } from "./category-create-request-dto.type.js";
 export { type CategoryGetAllItemResponseDto } from "./category-get-all-item-response-dto.type.js";
+export { type CategoryGetAllQuery } from "./category-get-all-query.type.js";